+++ lib/src/method_channel/mobile_scanner_method_channel.dart
@@ -501,7 +501,6 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
   }
 
   @override
-  @override
   Future<Set<CameraLensType>> getSupportedLenses({
     CameraFacing? facing,
   }) async {