+++ test/mobile_scanner_controller/get_supported_lenses_test.dart
@@ -77,8 +77,7 @@ class FakeMobileScannerPlatform extends MobileScannerPlatform {
   }
 }
 
-class FakeMobileScannerPlatformWithFacingCapture
-    extends MobileScannerPlatform {
+class FakeMobileScannerPlatformWithFacingCapture extends MobileScannerPlatform {
   FakeMobileScannerPlatformWithFacingCapture(
     Set<CameraLensType> supportedLenses,
   ) : _supportedLenses = supportedLenses;